+++ lib/td_dd/search/indexer.ex
@@ -5,7 +5,6 @@ defmodule TdDd.Search.Indexer do
 
   alias Elasticsearch.Index
   alias Elasticsearch.Index.Bulk
-  alias Jason, as: JSON
   alias TdCache.Redix
   alias TdDd.DataStructures.DataStructureVersion
   alias TdDd.Search.Cluster
@@ -20,7 +19,7 @@ defmodule TdDd.Search.Indexer do
   def reindex(:all) do
     Mappings.get_mappings()
     |> Map.put(:index_patterns, "#{@index}-*")
-    |> JSON.encode!()
+    |> Jason.encode!()
     |> put_template(@index)
     |> case do
       {:ok, _} ->